# app/07-swagger/javascript-client-generated/src/model/Response.js
/**
 * Ticket API
 * Ticket API for illustrating purpose.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: jlguenego@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 */


import ApiClient from '../ApiClient';





/**
* The Response model module.
* @module model/Response
* @version 1.0.0
*/
export default class Response {
    /**
    * Constructs a new <code>Response</code>.
    * @alias module:model/Response
    * @class
    */

    constructor() {
        

        
        

        

        
    }

    /**
    * Constructs a <code>Response</code> from a plain JavaScript object, optionally creating a new instance.
    * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
    * @param {Object} data The plain JavaScript object bearing properties of interest.
    * @param {module:model/Response} obj Optional instance to populate.
    * @return {module:model/Response} The populated <code>Response</code> instance.
    */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new Response();

            
            
            

            if (data.hasOwnProperty('content')) {
                obj['content'] = ApiClient.convertToType(data['content'], Object);
            }
            if (data.hasOwnProperty('error')) {
                obj['error'] = ApiClient.convertToType(data['error'], Object);
            }
        }
        return obj;
    }

    /**
    * @member {Object} content
    */
    content = undefined;
    /**
    * @member {Object} error
    */
    error = undefined;








}