# app/07-swagger/javascript-client-generated/src/model/Ticket.js
/**
 * Ticket API
 * Ticket API for illustrating purpose.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: jlguenego@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 */


import ApiClient from '../ApiClient';





/**
* The Ticket model module.
* @module model/Ticket
* @version 1.0.0
*/
export default class Ticket {
    /**
    * Constructs a new <code>Ticket</code>.
    * @alias module:model/Ticket
    * @class
    */

    constructor() {
        

        
        

        

        
    }

    /**
    * Constructs a <code>Ticket</code> from a plain JavaScript object, optionally creating a new instance.
    * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
    * @param {Object} data The plain JavaScript object bearing properties of interest.
    * @param {module:model/Ticket} obj Optional instance to populate.
    * @return {module:model/Ticket} The populated <code>Ticket</code> instance.
    */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new Ticket();

            
            
            

            if (data.hasOwnProperty('id')) {
                obj['id'] = ApiClient.convertToType(data['id'], 'String');
            }
            if (data.hasOwnProperty('name')) {
                obj['name'] = ApiClient.convertToType(data['name'], 'String');
            }
        }
        return obj;
    }

    /**
    * @member {String} id
    */
    id = undefined;
    /**
    * @member {String} name
    */
    name = undefined;








}